* Fixed #1954 (Locking issues between backend and frontend)
From now on, the backend is properly locking, so if two instances
are run at the same time, the second one will wait until the first
one has completed it's operations, and continue as usual thereafter.

Thus, most of the GUI lock checks that had to be done are now 
unnecessary, so I've removed them. The exception is with scripts
that add/modify multiple object types at the same time (objects
managed by different managers).

For example, the domain addition script also configures some default
htaccess permissions. Therefore, without locking it would be possible
that the following happens:

1. rqst-mngr (1st instance) is running
2. gui adds a new domain
3. rqst-mngr (2nd instance) runs, but waits blocked
4. rqst-mngr (1) launches htacces mngr, who tries to add the .htaccess
   related files, failing miserably because the domain (and it's
   folders) have not been created yet.

Obviously, this is not a problem when the script only modifies one
object type, such as:

1. rqst-mngr (1) is running
2. gui adds mail account (you can only do this for "ok" domains!)
3. rqst-mngr (2) runs, but waits blocked
4. rqst-mngr (1) launches mbox-mngr, who adds the mail account
5. rqst-mngr (1) ends
6. rqst-mngr (2) continues operation, but has nothing to do. No
   problem, just end.



git-svn-id: https://i-mscp.svn.sourceforge.net/svnroot/i-mscp/trunk@1952 15a38060-5cfe-4f5c-bf9d-b5d42b1378a7